Fixed a security issue in "domains" action when deleting domains + databases
A malicious user was able to trick the script into removing a completely
different and unrelated MySQL database (e.g. 'mysql' or another client). This
was possible because the stored ".ozsrvutils-mysql" file which contained the
username and password was editable by the user.

If this user had access to the file, e.g. FTP/SFTP/SCP then they could edit this
file and replace the database name with anything they wanted. The "domains"
action would then read this value and remove that database.

The ".ozsrvutils" and ".ozsrvutils" files are no longer generated, and the MySQL
details are now read from the credentials file only.